Task 3: Extract agent/session_revival.py — revival detection, cooldown tracking, stale branch cleanup
